[flutter] refactor keyboard and display library structure
test: manual, bots

Extensive use of source_dir ="." breaks code coverage as it combines the test and source files into the same package. This is a part of a global refactoring to move off of this pattern, as well as to make sure that dart packages within fuchsia conform to the common dart structure when possible.
For more information see: https://docs.google.com/document/d/1k3kOm4wJThEJIWIu5hYewiSmRGyBJd0Rzzx5_pQh78M/edit

Change-Id: I5e6a22d92859f555be3b75edac613cc61eaef6b4