[tests] Generates flutter driver extension wrapper
This change generates a flutter driver extension wrapper for modules
being run in debug mode.  This is a check (that can be disabled by
turning off assertions) for whether the module in question is running
inside of a testing environment (by having access to the TestRunner FIDL
service).

When built in release mode this code will not be run.

Re-lands from revert here:

Iac45be204b4891ac987d085052d97c58fd59878b

With the addition of:
```
if (defined(invoker.flutter_driver_extendable)) {
  not_needed(invoker, ["flutter_driver_extendable",])
}
```
For release builds built w/ AOT.

TEST: Running topaz modular integration tests. CQ/CI. Building w/ AOT forced
Change-Id: Ic48c0fba161897d1530bde1f1d157451c35be9f6