[tests] Generates flutter driver extension wrapper
This change generates a flutter driver extension wrapper for modules
being run in debug mode.  This is a check (that can be disabled by
turning off assertions) for whether the module in question is running
inside of a testing environment (by having access to the TestRunner FIDL
service).

When built in release mode this code will not be run.

TEST: Running topaz modular integration tests. CQ/CI
Change-Id: I5cae7eb7ccc15b179f42bb172fcc9a8c6ef4408f