[web_view] Cancel web focus requests when url changes
Previously, there was a bug where the keyboard would stick around if a
text input was focused at the moment that the URL changed. (Think
pressing the return key — tapping a submit button wouldn't work, since
it would defocus the input a moment before actually submitting the
form.)

Not really sure how this works with various sorts of AJAX requests, or
if it's the best solution for this bug, but it at least fixes the bug in
question? Seems better to err on the side of accidentally closing the
keyboard instead of accidentally keeping it open, at least at this point
in time?

Test: manually ensured keyboard closed correctly

Change-Id: If0ac78beadbf3cd06a49c8afc5b1a562ce5f7637