[components][flutter][dart] Propagate package_root
dart_library assumes package_root to be ".", the same path to the
BUILD.gn that defines the flutter/dart apps. It then proceeds to locate
source files by appending suffixes such as /lib.

This is not acceptable in the >1cpp world, as:
1. BUILD.gn is guaranteed to be in a different path for at least one
component.
2. components need to explicitly specify themselves.

This parameter is also better served as component_root, but that's
another CL for another time involving //build changes.

TESTED=echo_dart, simple_flutter builds

CP-142 #comment

Change-Id: I4761ecc5973123d6668d3d7d6ac69e14dbacae65