[jiri] Increase history depth to 50 for third party repos.
This will reduce the probability of a jiri update failing during a
release build. Currently those builds are failing if they refer to a
commit that is anything other than the most recent commit in that repo.
These repos have very few commits going into them, and we don't
often try to build old release snapshots, so this should in practice
eliminate the error.

This is a temporary fix. In the near term we will probably eliminate
all of these problematic repos. In the long term, we should replace
historydepth with shallow-exclude.

Test: not needed
Bug: INTK-485 #comment
Change-Id: I94703330ac3e1ecd44d09f421c5f443336b72f94