[sledge] Introduce notion of field paths.
To have a convenient querying API in Sledge, the fields need to be
addressable using a string.
Because schemas can embed other schemas, this string is a "field path"
composed of field names concatenated with the '.' separator.

This CL adds a method to check whether a field path points to a field
that exists in a schema.
This CL also makes it incorrect to have field names contain '.'.

TEST=The new functionalities are tested.

Change-Id: I95244fe490a59cd8bf41aa1f93ee1debd6755c79