[cmx] Add TestRunner to slider_mod_tests.cmx
https://fuchsia-review.googlesource.com/c/topaz/+/219808 added
"services" to this whitelist, and passed CQ in the dry run. But
then https://fuchsia-review.googlesource.com/c/topaz/+/218305
was submitted, which broke 219808. 219808 landed anyway because the
tryjobs from the previous dry run were green.

To fix, we just need to add TestRunner to the cmx.

TESTED=run_topaz_modular_integration_tests.sh

Change-Id: Iaf9adf14971093ddf882058a9ea247db94b750dd