[sledge] Add class storing queries.
The Query class stores the data required to do queries.

The final implementation of `prefixInIndex` is blocked on having a list
of all the indexes configured.

`getDocumentIds` now takes a `query`. A query can be created with just
a schema, in which case `getDocumentIds` works the same way as before.

Note:
A `QueryBuilder` class will later be introduced that will offer a
convenient way of creating Queries. Client won't be expected to use
the Query constructor themselves.

test=CQ + new test added.

Change-Id: I44a963acc8e45f15be6ae0231f6c89e525453949