[dart-sdk] Initial impl of Agent
Provides a new API that takes care of initializing an Agent,
exposing it to the rest of the system and registering to
life-cycle events. Additionally, provides a temporary
implementation of addServices method.

Also, refactored the Fibonacci Agent example to illustrate
the reduction of boiler plate code and in order to validate
that the code works in action.

Test: fx run-host-tests fuchsia_modular_package_unittests

Change-Id: I7dbcde27fffc8c6e8e84710d6b95f77fe4694b24