[dart-sdk] Add exposeSerivce API to Agent
This provides a convince method to associate
a service implementation to the Agent and exposes
it to the rest of the system so that it can be
discovered and connected to.

As outlined in the Module SDK design, the FIDL
Dart bindings had to be changed in order to
to make this possible.

Test:
1. fx run-host-tests fuchsia_modular_package_unittests
2. fx shell /pkgfs/packages/topaz_modular_integration_tests/0/bin/run_topaz_modular_integration_tests.sh slider_mod_tests

Change-Id: I4cadbcd21d401d39ead71d5e7b154c66ebe217f7