[settings][bt]Fix sorting in bluetooth settings
Sort is supposed to return a positive number if a is greater, zero if
equal, or a negative number if smaller. However, a missing set of
parens caused the lambda to always return a negative number (since the
rssi is always a negative number), meaning no actual sorting was done.

SY-879
Test: ran it and verified sort was correct
Change-Id: I196e90b1a94cb88605db3f41d87e24322ec63028