[sledge] Add a QueryBuilder class.
NumFieldValue supports both doubles and integers.
The reason is that traditionally queries don't differentiate between
integers and doubles.
For example, a query with "a > 42" will work regardless of whether [a]
contains integers or doubles.
(As a reminder, *FieldValues are used in Queries to hold the values used
in equalities and inequalities.)

TEST=dart_sledge_tests

Change-Id: Ia0c030127fc0326a9cc13ac85cbc884a612846c4