[sledge] Add delete method to the Document class.
Document now have a `delete` method that is callable in transactions.
When called, the `delete` method results in sledge removing all the
key-values backing the document, unless the transaction is aborted.
Once deleted, access to the document results in exceptions being thrown.

In follow up CLs, a custom conflict strategy will be added to correctly
merge deletions, and Sledge's API will be changed to make it
impossible to recreate documents.

TEST=dart_sledge_tests

Change-Id: I2fc28231c156f0cf8a46fc32f72ddbcc4dc61e7a