[sledge] Improve Sledge initialization.
Callbacks passed to FIDL bindings are executed with a different stack
than the one that initiated the call to the FIDL binding.
This means that you can't catch exceptions thrown from inside callbacks
executed by FIDL bindings.

With this CL, errors that happen during the initialization of Sledge do
not throw exceptions anymore.
Instead, the presence of error is only recorded in the
initializationCompleter Completer.

TEST=ran sledge_testing_mod

Change-Id: Ifb131d639185380254a4109468276bde2f732eb0