[chromium] Give the chromium enablement GN flag a more sensible name
This renames 'enable_chromium_based_web_runner' to just
'enable_chromium' since it now controls more than just the web_runner
implementation, and is a bit of a mouthful.

This leaves the previous spelling of the GN arg in place and defaults
it to the same as 'enable_chromium' to enable a soft transition with
code in other repositories that refers to this value.

Test:
US-472 #comment

Change-Id: Ib1c6735ed6a4a4e7b4744f0646e2840c7939a9cf