[prebuilt-dart-sdk] Change CIPD root to be local to script
CIPD has a notion of a CIPD root where it stores its internal cache of
files and metadata. The 'cipd ensure' command assumes there is exactly
one ensure file per root and removes all cipd-managed entries that are
not in that ensure file in that root.  We'd like to use the root of the
Fuchsia checkout for <package> entries in jiri manifests.

This modifies the tools/prebuilt-dart-sdk/update.py script to use
tools/prebuilt-dart-sdk/ as its CIPD root instead of the root of the
Fuchsia tree.  Our other isolated 'cipd ensure' invocations, like in
https://fuchsia.googlesource.com/garnet/+/master/tools/cipd-update.sh
and
https://fuchsia.googlesource.com/topaz/+/master/tools/cipd-update.sh
use more local cipd roots. This script is a bit unusual compared to our
other cipd users since it computes the contents of the ensure file
dynamically.

IN-767 #comment move prebuilt-dart-sdk cipd root
Test: CQ

Change-Id: I1c5502121f2e9d89bdce06b97e2dab5a97cc775e