[flutter] Minor tracing adjustments
Convert all `TRACE_EVENT*`s to `TRACE_DURATION`s, which have the benefit of
being able to automatically detect arity.

Add a size argument to the "VulkanSurfacePool::CreateSurface" trace.

Move the "VulkanSurface::AllocateDeviceMemory" trace to "vk().AllocateMemory",
which is the section of it that dominates time.  Add a size argument to it as
well.

Test: Taking and inspecting a trace + CQ

Change-Id: Ib2d7267139b0015af9bd484105c512db5d4aa47a