[tracing] Move tracing arguments out of name into arguments
Move arguments out of the name of the trace event, into arguments
instead.  This has the advantage of making it easier to group events of
the same type together, in both the trace viewer and external processing
scripts.

In doing so, switch from using the |trace| API to using |Timeline|
directly.  See Ic74ee23932972b1d33708cc74575f50dc4dd9d9d for further
context on this decision.

Test: CQ
Change-Id: Icb798f5439f963d28e50f8a55bb8261f5c1b46d7