[runtime][flutter_runner] Handle non-linear image tiling correctly
Tiling::LINEAR was incorrectly passed to Scenic for tiled images.

This happen to work as Scenic ignores this value and
assumes that Tiling::GPU_OPTIMAL is always used. In order to
have Scenic support Tiling::LINEAR we need flutter runner to
pass image info to Scenic with correct tiling value.

Use VkImageCreateInfo tiling value to determine the tiling
parameter to pass to Scenic.

Test: linear image support in Scenic doesn't break flutter runner
Change-Id: Ib4b0575c70c3b0ca90d3c29aa6597add177a58f5