[deprecated_loop] Support replacing std::function with fit::function
Removing message_loop_unittest case:
"TaskRunnerAvailableDuringLoopDestruction"

With the new move-only semantics of fxl::Closure, this test no longer
passes. The test only compiles if the test inner class
"DestructorObserver" is copyable. If I make DestructorObserver a
move-only class, or wrap in "make_unique<>", the assignments to
"d" (which was in the original test case) fail to compile. This seems
to indicate that the assignments are copying the DestructorObserver.

Without making the copies, the destructor doesn't invoke the
DestructorObserver callback and the tests fail. It's possible that
the destructor was invoked for the original instance and not the
instance assigned to "d", which was added to the task_runner.
The test may not have been testing the intended behavior in the first
place. (I may be wrong.)

Since this "MessageLoop" class is deprecated anyway, and all other
tests pass, I'm hoping this behavior is not relevant now, if it was
ever relevant in the past.

Test: message_loop_unittest
Change-Id: I583e3824e694050158c352b59d8a0a5ae6961ca8