[topaz] 2nd of 3-phase topaz change to remove MakeCopyable
This change requires change https://fuchsia-review.googlesource.com/c/garnet/+/240405
to be merged first.

Note: The above garnet change includes the first change to topaz (1st
change), this is the second topaz change, and should compile cleanly.
But we still need to change lines 95-96, which depends on a future
change to peridot: modular::testing::Done will change from taking
a std::function to taking a fit::function, and "done" will need to be
wrapped with std::move().

Test: No behavior change
Change-Id: I3d073acec5381c12e48d3acc183e4895577a0b2c