[storyshell] Add Mondrian library as part of story shell refactor
This commit is an implementation of the API and a placeholder
implementation of the minimal viable logic to unblock Presenter work.

Background: Step 1 of separating story shell into (this) layout library,
a UI & interaction concerned presenter, and a component that provides
the story shell FIDL API, for increased flexibility.

Adds tests for library functionality:
  surface_node_test, layout_test, surface_tree_test, encode_decode_test

Adds this library's tests to dart_unittests

Test: with fx set x64 --board x64 --available=topaz/packages/tests/all
then fx run-host-tests mondrian_lib_tests

Change-Id: Ia680636aed6d2c840cb861392ee8902749f376fd