[web_runner][test] Input x Chromium x Scenic test
This test loads a dynamic page that starts with a Fuchsia background and has a
large text box, with Javascript to change the background to the color typed
in the text box. This test verifies the initial color, taps on the text box
(top quarter of the screen), types a new color, and verifies the changed
color.

SCN-1260 #done
Test: fx run-test web_runner_tests

Change-Id: I42427f6dfa6ae77eaeb0d5131bad746e7dcced66