[child_view] Commonize ViewContainer in the SDK
Prepare for moving all Childview use to the SDK by commonizing some code
first.

This code creates a singleton that roughly maps to the ViewProvider
interface that a mod exposes.  Unfortunately it causes problems when a
mod is using both the SDK and non-SDK versions of child_view.dart,
because the 2nd singleton initialization fails.

So, we use only the SDK singleton and make the non-SDK code rely on it.

TEST: ran SysUI and launched a story; asked to play a movie trailer
SCN-898 #comment
SCN-1214 #comment

Change-Id: Iae21710a2a20c9d97e6f1a201ef69b2f72f9dec8