[persistent-storage] Rename feature to 'deprecated-global-persistent-storage'
This modifies all component manifests in this repo that previously
included the 'persistent-storage' sandbox feature to instead include
'deprecated-global-persistent-storage', which is the new name for the
same behavior. There is no functional impact of this change.

Components migrated here were included in the build-time whitelist in
fxr/254331

CF-28 #progress

Test: CQ, aka successful build (whitelist wasn't tripped) and no
regressions in behavior.

Change-Id: Id668110575a47c9f437fc51c4f47d458d8480dfc